refactor(java): organize root package by feature
