+++ sdks/java/src/main/java/org/byteveda/taskito/cli/Cli.java
@@ -4,13 +4,13 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CountDownLatch;
-import org.byteveda.taskito.DeadJob;
-import org.byteveda.taskito.Job;
-import org.byteveda.taskito.JobFilter;
-import org.byteveda.taskito.JobStatus;
 import org.byteveda.taskito.Queue;
 import org.byteveda.taskito.Taskito;
 import org.byteveda.taskito.dashboard.DashboardServer;
+import org.byteveda.taskito.model.DeadJob;
+import org.byteveda.taskito.model.Job;
+import org.byteveda.taskito.model.JobFilter;
+import org.byteveda.taskito.model.JobStatus;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
