+++ sdks/java/src/main/java/org/byteveda/taskito/locks/LockInfo.java
@@ -1,4 +1,4 @@
-package org.byteveda.taskito;
+package org.byteveda.taskito.locks;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
