+++ sdks/java/src/main/java/org/byteveda/taskito/locks/package-info.java
@@ -1,0 +1,2 @@
+/** Distributed, TTL-bounded advisory locks. */
+package org.byteveda.taskito.locks;
