+++ sdks/java/src/main/java/org/byteveda/taskito/model/JobError.java
@@ -1,4 +1,4 @@
-package org.byteveda.taskito;
+package org.byteveda.taskito.model;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
