+++ sdks/java/src/main/java/org/byteveda/taskito/model/package-info.java
@@ -1,0 +1,2 @@
+/** Immutable, read-only data the API returns: jobs, stats, dead letters, metrics, logs. */
+package org.byteveda.taskito.model;
