+++ sdks/java/src/main/java/org/byteveda/taskito/package-info.java
@@ -1,11 +1,20 @@
 /**
  * Taskito Java SDK — a typed client over the Rust task-queue core.
  *
- * <p>Open a {@link org.byteveda.taskito.Queue} via
- * {@link org.byteveda.taskito.Taskito#builder()}. Enqueue typed payloads with a
- * {@link org.byteveda.taskito.Task}; payloads serialize through a
- * {@link org.byteveda.taskito.serialization.Serializer}. The implementation
- * talks to the core through the {@link org.byteveda.taskito.spi.QueueBackend}
- * seam, whose default implementation lives in {@code internal}.
+ * <p>This root package is the front door: open a
+ * {@link org.byteveda.taskito.Queue} via
+ * {@link org.byteveda.taskito.Taskito#builder()}. Everything else is grouped by
+ * feature:
+ *
+ * <ul>
+ *   <li>{@code task} — {@link org.byteveda.taskito.task.Task} descriptors,
+ *       handler functions, enqueue options
+ *   <li>{@code model} — immutable views the API returns (jobs, stats, metrics)
+ *   <li>{@code worker} — the worker runtime
+ *   <li>{@code locks} / {@code scheduling} / {@code workflows} — feature surfaces
+ *   <li>{@code serialization} / {@code middleware} / {@code events} — cross-cutting
+ *   <li>{@code spi} — the {@link org.byteveda.taskito.spi.QueueBackend} seam,
+ *       whose default implementation lives in {@code internal}
+ * </ul>
  */
 package org.byteveda.taskito;
