+++ sdks/java/src/main/java/org/byteveda/taskito/scheduling/package-info.java
@@ -1,0 +1,2 @@
+/** Periodic (cron-scheduled) task registration. */
+package org.byteveda.taskito.scheduling;
