+++ sdks/java/src/main/java/org/byteveda/taskito/task/package-info.java
@@ -1,0 +1,2 @@
+/** Defining and shaping work: task descriptors, handler functions, enqueue options. */
+package org.byteveda.taskito.task;
