+++ sdks/java/src/main/java/org/byteveda/taskito/worker/WorkerDispatchBridge.java
@@ -1,4 +1,4 @@
-package org.byteveda.taskito;
+package org.byteveda.taskito.worker;
 
 import java.util.List;
 import java.util.Map;
