+++ sdks/java/src/main/java/org/byteveda/taskito/worker/package-info.java
@@ -1,0 +1,2 @@
+/** The worker runtime: handler registration, dispatch bridge, lifecycle. */
+package org.byteveda.taskito.worker;
