+++ sdks/java/src/test/java/org/byteveda/taskito/DashboardTest.java
@@ -11,6 +11,8 @@ import java.nio.file.Path;
 import java.util.Collections;
 import java.util.Map;
 import org.byteveda.taskito.dashboard.DashboardServer;
+import org.byteveda.taskito.task.EnqueueOptions;
+import org.byteveda.taskito.task.Task;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.io.TempDir;
