+++ sdks/java/src/test/java/org/byteveda/taskito/LockTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.nio.file.Path;
+import org.byteveda.taskito.locks.Lock;
+import org.byteveda.taskito.scheduling.PeriodicTask;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
