+++ sdks/java/src/test/java/org/byteveda/taskito/WorkerTest.java
@@ -11,6 +11,8 @@ import java.util.Optional;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import org.byteveda.taskito.events.EventName;
+import org.byteveda.taskito.task.Task;
+import org.byteveda.taskito.worker.Worker;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.io.TempDir;
